// packages/plugin-qti-interactions/src/schema/prosemirror-schema.ts
export const nodes: Record<string, NodeSpec> = {
  doc: { content: 'block+' },
  text: { group: 'inline' },
  paragraph: defineParagraph(),
  'qti_choice_interaction': {
    attrs: {
  'disabled': { default: null },
  'enabled': { default: null },
  'maxChoices': { default: null },
  'minChoices': { default: null },
  'readonly': { default: null },
  'required': { default: null },
  'responseIdentifier': { default: null },
  'shuffle': { default: null }
},
    parseDOM: [{ tag: 'qti-choice-interaction', getAttrs: (dom) => parseDomAttrs(dom, {
    'disabled': 'disabled',
    'enabled': 'enabled',
    'maxChoices': 'max-choices',
    'minChoices': 'min-choices',
    'readonly': 'readonly',
    'required': 'required',
    'responseIdentifier': 'response-identifier',
    'shuffle': 'shuffle'
  }) }],
    toDOM: createToDOM('qti-choice-interaction', true, false, {
    'disabled': 'disabled',
    'enabled': 'enabled',
    'maxChoices': 'max-choices',
    'minChoices': 'min-choices',
    'readonly': 'readonly',
    'required': 'required',
    'responseIdentifier': 'response-identifier',
    'shuffle': 'shuffle'
  }),
    group: 'block',
    content: 'qti_prompt? qti_simple_choice+',
    defining: true,
    isolating: true,
  },
  'qti_prompt': {
    attrs: {},
    parseDOM: [{ tag: 'qti-prompt', getAttrs: (dom) => parseDomAttrs(dom, {}) }],
    toDOM: createToDOM('qti-prompt', true, false, {}),
    group: 'qti',
    content: 'block+',
    defining: true,
  },
  'qti_simple_choice': {
    attrs: {
  'disabled': { default: null },
  'fixed': { default: null },
  'identifier': { default: null },
  'readonly': { default: null },
  'selected': { default: null }
},
    parseDOM: [{ tag: 'qti-simple-choice', getAttrs: (dom) => parseDomAttrs(dom, {
    'disabled': 'disabled',
    'fixed': 'fixed',
    'identifier': 'identifier',
    'readonly': 'readonly',
    'selected': 'selected'
  }) }],
    toDOM: createToDOM('qti-simple-choice', true, false, {
    'disabled': 'disabled',
    'fixed': 'fixed',
    'identifier': 'identifier',
    'readonly': 'readonly',
    'selected': 'selected'
  }),
    group: 'qti',
    content: 'paragraph+',
    marks: '_',
    defining: true,
  },
  'qti_text_entry_interaction': {
    attrs: {
  'responseIdentifier': { default: null }
},
    parseDOM: [{ tag: 'qti-text-entry-interaction', getAttrs: (dom) => parseDomAttrs(dom, {
    'responseIdentifier': 'response-identifier'
  }) }],
    toDOM: createToDOM('qti-text-entry-interaction', false, true, {
    'responseIdentifier': 'response-identifier'
  }),
    inline: true,
    group: 'inline',
    marks: '_',
    atom: true,
    selectable: true,
  },
};

export const marks: Record<string, MarkSpec> = {};